05-agent-fw/hosted-agent/requirements.txt
azure-ai-agentserver-agentframework==1.0.0b7
agent-framework-azure-ai==1.0.0b251211
